Use abstract class for tables/datasets tabs
